moved Header component to components folder and rename existing <Header/> imports
